Add files via upload
Fertigation is related to chemigation, the injection of chemicals into an irrigation system. The two terms are sometimes used interchangeably however chemigation is generally a more controlled and regulated process due to the nature of the chemicals used. Chemigation often involves insecticides herbicides, and fungicides, some of which pose health threat to humans, animals, and the environment.